Testing 20240501

+++ Backend/src/repository/precioRespository.ts
@@ -20,13 +20,12 @@ export class PrecioRepository  {
     }
     
 
-    async findOne(item: { fechaDesde: Date, componenteId:number }): Promise<Precio | undefined> {
+    async findOne(item: { id:number }): Promise<Precio | undefined> {
         try {
             
             const precio = await em.findOneOrFail(
                 Precio,
-                { fechaDesde: item.fechaDesde, 
-                     componente:{id: item.componenteId}
+                { id: item.id
                  }
                  ,{ populate: ['componente'] }
             );
@@ -53,8 +52,10 @@ export class PrecioRepository  {
         try {            
 
             if(!item.componente.id){return undefined;}
+            if(!item.id){return undefined;}
             
-            const precioToUpdate = await this.findOne({fechaDesde:item.fechaDesde, componenteId:item.componente.id })
+            
+            const precioToUpdate = await this.findOne({id:item.id})
             if(precioToUpdate){
             em.assign(precioToUpdate, item)
             await em.flush()
@@ -70,11 +71,11 @@ export class PrecioRepository  {
     }
     
 
-    async delete(item: { fechaDesde: Date, componenteId:number }): Promise<Precio | undefined> {
+    async delete(item: { id:number }): Promise<Precio | undefined> {
         try {
             
-              
-            const precio = await this.findOne(item);
+            if(!item.id){return undefined;}
+            const precio = await this.findOne({id:item.id});
 
             if (precio) {
                 await em.removeAndFlush(precio);
@@ -96,11 +97,11 @@ export class PrecioRepository  {
     async updateValor(item: Precio, newValor:number): Promise<Precio | undefined> {
         try {
             
-            if (!item.fechaDesde || !item.componente.id) {
+            if (!item.fechaDesde || !item.componente.id||!item.id) {
                 console.error('ERROR');
                 return undefined;
             }
-            const precioToUpdate = await this.findOne({ fechaDesde: item.fechaDesde, componenteId: item.componente.id });
+            const precioToUpdate = await this.findOne({ id:item.id });
             if (precioToUpdate) {
             precioToUpdate.valor = newValor;
             await em.persistAndFlush(precioToUpdate);
